+++ src/Components/Routing.js
@@ -9,10 +9,8 @@ import UserAuth from "./UserAuth";
 import LogMeal from "./LogMeal";
 import Hist from "./Hist";
 import Analysis from "./Analysis/Analysis";
-import AnalysisDay from "./Analysis/AnalysisDay";
 import AnalysisWeek from "./Analysis/AnalysisWeek";
 import * as Items from "./Data/DummyData";
-import WeeklyCal from "./Analysis/WeeklyCal";
 import WeeklyCarbo from "./Analysis/WeeklyCarbo";
 
 export default function Routing({ loggedInUser }) {
@@ -24,14 +22,9 @@ export default function Routing({ loggedInUser }) {
       <Route path="/" element={<Home logInUser={loggedInUser} />} />
       <Route path="/about" element={<About />} />
       <Route path="/faq" element={<FAQ />} />
-      <Route path="/analysis" element={<Analysis />}>
-        <Route path="daily" element={<AnalysisDay data={dummyData} />} />
+      <Route path="/analysis" element={<Analysis data={dummyData} />}>
         <Route path="weekly" element={<AnalysisWeek data={weekData} />}>
-          <Route path="cal+sodium" element={<WeeklyCal data={weekData} />} />
-          <Route
-            path="carbo+others"
-            element={<WeeklyCarbo data={weekData} />}
-          />
+          <Route path="others" element={<WeeklyCarbo data={weekData} />} />
         </Route>
       </Route>
       <Route path="/login" element={<UserAuth />} />
